+++ component/phlare/scrape/scrape.go
@@ -72,7 +72,7 @@ type Arguments struct {
 	// // scrape to fail.
 	// LabelValueLengthLimit uint `river:"label_value_length_limit,attr,optional"`
 
-	HTTPClientConfig component_config.HTTPClientConfig `river:"http_client_config,block,optional"`
+	HTTPClientConfig component_config.HTTPClientConfig `river:",squash"`
 
 	ProfilingConfig ProfilingConfig `river:"profiling_config,block,optional"`
 }
@@ -203,7 +203,8 @@ func (arg *Arguments) UnmarshalRiver(f func(interface{}) error) error {
 		}
 	}
 
-	return nil
+	// We must explicitly Validate because HTTPClientConfig is squashed and it won't run otherwise
+	return arg.HTTPClientConfig.Validate()
 }
 
 // Component implements the pprof.scrape component.